user32/tests: Add ClipCursor / focus interactions tests.
This is to confirm the behaviour of SetForegroundWindow and other focus
switching methods w.r.t. the active ClipCursor rect.

Depending on the method used to switch windows, the ClipCursor rect may
be reset or not. Alt-Tab seems to reset it most of the time, but not on
all windows versions. Other methods like closing windows, or calling
SetForegroundWindow usually do not.

MSDN says the applications should release the cursor when losing the
focus, and, as windows doesn't seem very consistent, we will assume that
they do.