user32/tests: Add concurrency tests for SetForegroundWindow.
When calling SetForegroundWindow for a window in another thread, an
internal message is posted to the thread's message queue.

If this thread then calls SetForegroundWindow before processing its
messages it will execute the corresponding set_active_window first,
but then overwrite the active window later, when processing its internal
messages.

This is not always the correct behavior and these tests help determine
what should actually be done in various situations.

This aims to check the following sequences, with A being a separate
thread that created three windows, and B being the main test thread
with some windows initially in background:

* window A0, A1, or A2 is foreground, then:
  * B sets foreground to window A0
  * A sets foreground to window A1

As well as these sequences where foreground is also temporarily switched
to window B0:

* window A0, A1, or A2 is foreground, then:
  * B sets foreground to window B0
  * B sets foreground to window A0
  * B sets foreground to window B0
  * A sets foreground to window A1

In addition, we also do tests with additional SetActiveWindow / SetFocus
calls to check their influence.

user32/tests: Add SetForegroundWindow message sequence tests.