wined3d: Add swapchain parameter to wined3d_swapchain_set_fullscreen.
We need that in the following patches to access swapchain_ops.

This parameter may be NULL if the swapchain is not actually a wined3d
swapchain (in the case of d3d12 swapchain).

For d3d11 swapchain, we could pass the wined3d_swapchain but dxgi
requires some additional window style changes that will be added in a
separate patch.