server: Support mounting a FUSE overlayfs on drive_c.
This uses fuse-overlayfs [1] to stack layers under drive_c and save some
space. It is then possible to have a default wine prefix that will be
used as a read-only base for other wine prefixes, with only the new or
modified files saved into the upper dir.

The overlayfs is automatically mounted by the server on startup, and
unmount is requested when the server exits. If fuse-overlayfs fails for
any reason, it should fallback to normal prefix and the upper dir will
be updated.

The lower dirs are controlled by the WINEPREFIX_OVERLAYS environment
variable, which is passed as is after -olowerdir= to fuse-overlayfs. It
supports multiple layers, separated by a colon, from top to bottom.

Any other option supported by fuse-overlayfs can also be used and passed
after the lower dirs in WINEPREFIX_OVERLAYS, separated from them with
comma - for example uidmapping/gidmapping.

[1] https://github.com/containers/fuse-overlayfs