server: Remove non-responsive hooks, as Win7 does.
GTA V installs a keyboard hook on startup but then never checks its
messages. The user side already removes the hook on first timeout, but
if there's no other user-side hook, the server is always going to wait.

Usually the dinput mouse/keyboard hook are removing other non-responsive
hooks, but in order to be able to stop using hooks in rawinput mode we
have to support removal on timeout on the server side as well.