ntdll: Implement debugstr_a/w(n) format extensions.
This makes it possible to write TRACE("%p(astr)", "string") and get the
equivalent of TRACE("%s", debugstr_a("string")), respectively with
"%p(wstr)" and debugstr_w(...).

The width format specifier (including '*') can be used to control the
length (GCC format checker doesn't like precision specifier on %p).

Note that the format extension is just here to illustrate, we could
think of any other syntax. The main issue is to pass GCC format checker,
and as done in the Linux kernel %p suffixes seems to be the best fit.