wineboot: Compute and write the TSC frequency to registry ~Mhz.
In HKEY_LOCAL_MACHINE\\HARDWARE\\DESCRIPTION\\System\\CentralProcessor

Squashed with patches from:

* Arkadiusz Hiler <ahiler@codeweavers.com>

Check if the kernel trusts TSC before using it for Qpc.

Even if the bits are claiming that TSC meets our requirements the
hardware implementation may still be broken.

The Linux kernel does a lot of quality testing before deciding to use as
the clock source. If it (or the user, through an override) does not trust
the TSC we should not trust it either.

* Joshua Ashton <joshua@froggi.es>

Some games such as Horizon Zero Dawn use this registry value to
correlate values from rtdsc to real time.

Testing across a few devices, is seems like Windows always returns the
TSC frequency in this entry, not the current/maximum frequency of the
processor.

Returning the nominal/maximum cpu frequency here causes the game to run
in slow motion as it may not match the tsc frequency of the processor.

Ideally we'd not have to measure this and the kernel would return
tsc_khz to userspace, but this is a good enough stop-gap until
https://lkml.org/lkml/2020/12/31/72 or something similar is merged.