winepulse.drv: Register pulse_write as write callback.
This allows us to write more data to pulse when it notifies us, in
addition to the timed writes, which may not write anything if pulse
isn't ready.

This greatly reduces the amount of underflows in Prince of Persia: The
Forgotten Sands, which started clicking since we moved to PE Faudio.

The reason is not completely clear, but may be related to the game
creating a 6-channel master voice, even if the output is only stereo,
perhaps changing pulse buffer requirements and making it harder to catch
up missed writes.