xinput1_3: Write rumble and buzz reports asynchronously.
Some devices may block when writing output reports, and now that we can
write them in winebus.sys asynchronously, we need to also request async
operations on the user side.

This also waits for 1ms for the write to complete before cancelling the
previous writes, as we only track one pending write for each motor, to
leave a bit of time for the operation to take effect and not cancel
every write.