HACK: qasf: Only force user provided time to -1 if no quartz clock is set.
This satisfies both Sang Froid intro song which expects the audio stream
to fast forward decode by not providing any sync source to the filter,
and Space Engineers which expects the filter to respect the system clock
timing when decoding its videos.

This is likely a hack and there's probably something to do with wmvcore
sync clock source to make it match quartz clock.

CW-Bug-Id: #21139