configure.ac: Use ; rather than && to chain variable and comparison.
The variable assignment doesn't return success if git fails, and doesn't
link to the && command which compares the value. This is the case for
instance when building outside of a git tree, in which case the returned
build or revision is empty, and it forces a rebuild of winetest in that
case.