winegstreamer/wg_parser: Set n-threads for dav1d to 1.
It defaults to `$(nproc)` which causes the memory footprint to skyrocet
on high core count systems.

Setting it to 1 makes it use a bit less memory that theora decoder and
is enough to decode 1080p 30fps stream on 5950x with a speed of 4x the
real time.

We may want to set it to value higher than 1 for 64bit apps where
limited address space is not an issue. 2 or 4 may be optimal for tiling
with 2 columns and 2 rosw when encoding the video.

CW-Bug-Id: #19516