mf/tests: Test byte stream handler to media source resolution time.
Unreal Engine has a race condition in its media player where it starts
the resolution of a media source, waiting for its completion in an async
task, and expects to set the media player source from its result.

At the same time, the media player uses timer to tick every 200ms or so,
and if the first tick happens before the media source resolution end,
the media player ends up in a corrupted state and playback may fail or
get stuck.

On Windows, the media source resolution takes very little time, and
usually less than 1ms. On Wine, because of the current design of the
media soruce, it can often take more than 200ms depending on the setup,
and very rarely less than 50ms.