HACK: winegstreamer/wm_reader: Use OpenGL video processing pipeline for WMReader.
There's many 32-bit games using either the WM reader or the ASF reader
filter, and which require conversion from I420 to RGBA. However, ORC
fails to create an optimized routine as it doesn't support spilling
vector registers and it needs more than what's available in 32-bit x86.

The WM reader (and the ASF reader filter) currently reads audio and
video samples synchronously, and having a high video decoding latency
causes trouble on the audio stream.

Using OpenGL to post-process the video, and more specifically for the
color conversion makes it generally faster, enough to solve the problem
in most cases.

CW-Bug-Id: #18799
CW-Bug-Id: #18994
CW-Bug-Id: #20182
CW-Bug-Id: #20363
CW-Bug-Id: #20905
CW-Bug-Id: #20980