MR
This is a POC, showing how we could build Zink in Wine, to offer optional OpenGL implementation over Vulkan. The MR commits are split to show the required adjustments to build in-tree separately from the source import itself. There's also a couple of separate fixes which I've also sent as upstream merge requests but are not yet merged.

It has several advantages over an externally built module, but also several aspects that could be problematic. The main and biggest advantage of course is that it would work out of the box. Having an externally build module would require a more complicated setup to distribute it.

Another reason for doing this like that is that Mesa WGL isn't fully implemented. There's a couple of missing functions, some of which we have implemented on our side, and having the module builtin would let us more easily fill the gaps or integrate the code with our GDI implementation.

It also has some behavior differences with native opengl32, and although I could run some basic applications as well as several games with this MR, our opengl32 tests don't pass. Having the code in tree could make it easier to improve compatibility.

About the problematic aspects:

1) Mesa uses C++ code in a couple of places. There isn't a lot of code, and this MR shows that we could support it nonetheless with a little amount of plumbing, without relying on C++ headers or runtime which MSVC Clang doesn't have.

2) With only a couple of weeks of history it's not clear whether updating the source from upstream is going to be problematic. I only included the sources which were needed (on a logical module granularity level), and the Wine-specific adjustments to build in-tree are relatively small, but it's still unknown.

3) There's a large amount of generated code from python scripts. I've committed the generated files, as we often do, but maybe it would be better to have a way to generate them at build time. It's mostly just a concern about committed file size though.