winex11: Minimize windows that are outside of the virtual screen.
Unmapping the windows will make them disappear from the application
switcher, althought they are supposed to still be visible. Minimize
them instead, so that the user can switch back to them.

As we are requesting the minimization, the desired_state will match the
current_state when the corresponding PropertyNotify event will be
received, and we won't notify the Win32 side of the minimization. It
will also not be notified of window position changes as we don't send
them for minimized windows.

The Win32 state will stay as it is, until the windows are either moved
back on screen by the application, or switched to by the user. In the
latter case, a SC_RESTORE command is sent to the window.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57472
Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57474