winex11: Avoid creating windows with override-redirect flag set.
This confuses mutter and it manifests with spurious IconicState WM_STATE
change when it decides to try managing the window, but it also makes it
randomly lose focus or even fail to map the window back on screen.

Adding the flag after the window has been created fixes the issue.